requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
pycryptodome==3.9.6
PyJWT==1.7.1
python-weixin==0.5.0
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
xmltodict==0.12.0
